Add files via upload
This piece of code runs through the CASE 3 data of the Mosquito Habitat mapper on GLOBE observers and finds the number of reported habitats in three locations, Los Angeles, Charlotte, and Atlanta. The framework is pandas and geohash was imported to covert all the longitudes and latitudes into strings, therefore making it easy to count the number of cases in each area.